CNN Project/requirements.txt
fastapiuvicorntensorflownumpypillowpython-multipartjinja2